src/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
executing==2.1.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
grpcio==1.67.0
gymnasium==0.29.1
idna==3.10
imageio==2.36.0
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapy==1.2.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyboy==2.4.0
pygame==2.6.1
pygba==0.2.4
Pygments==2.18.0
pyparsing==3.2.0
PySDL2==0.9.16
pysdl2-dll==2.30.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
stable_baselines3==2.3.2
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tifffile==2024.9.20
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.162
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
